+++ lib/core/ui-basic/ui-basic-input/ui-basic-input.component.ts
@@ -20,4 +20,11 @@ export class UiBasicInputComponent extends CustomControlComponent {
   override registerOnChange(fn: any): void {
     this.onChange = fn;
   }
+
+  onInput(e: Event): void {
+    const value = (e.target as HTMLInputElement).value;
+
+    this.control.setValue(value);
+    this.onChange(value);
+  }
 }