+++ lib/ui-material/ui-material-input/ui-material-input.component.ts
@@ -37,4 +37,11 @@ export class UiMaterialInputComponent extends CustomControlComponent {
   override registerOnChange(fn: any): void {
     this.onChange = fn;
   }
+
+  onInput(e: Event): void {
+    const value = (e.target as HTMLInputElement).value;
+
+    this.control.setValue(value);
+    this.onChange(value);
+  }
 }